Include a new stub header to help with multilib compilation
Due to the unpredictable sort of includes, we'll often end up with
the 64-bit glibconfig.h header when building 32-bit packages, when
we need the one with 32-bit definitions.

To help alleviate that issue, we place a dynamic stub header into
the system include directory, which will automatically include the
right header location based on the word size.

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>